+++ cryptopedia-core/src/main/kotlin/me/hajoo/cryptopediaserver/core/domain/DailyVolumeStats.kt
@@ -25,15 +25,15 @@ class DailyVolumeStats(
     @Column(nullable = false)
     val date: LocalDate,
 
-    @Column(nullable = false, precision = 18, scale = 8)
+    @Column(nullable = false, precision = 32, scale = 8)
     val volumeSum: BigDecimal,
 
-    @Column(nullable = false, precision = 18, scale = 8)
+    @Column(nullable = false, precision = 32, scale = 8)
     val quoteVolumeSum: BigDecimal,
     
-    @Column(nullable = true, precision = 18, scale = 8)
+    @Column(nullable = true, precision = 32, scale = 8)
     val volumeMa7d: BigDecimal? = null, // Moving Average 7 Days
 
-    @Column(nullable = true, precision = 18, scale = 8)
+    @Column(nullable = true, precision = 32, scale = 8)
     val volumeMa30d: BigDecimal? = null
 )
